Use int instead of short for message lengths and counters.

+++ src/MStream.C
@@ -17,8 +17,8 @@ struct StreamMessage {
   char header[CmiMsgHeaderSizeBytes];
   int PE;
   int tag;
-  unsigned short len; // sizeof the data
-  unsigned short index; // index of packet in stream
+  unsigned int len; // sizeof the data
+  unsigned int index; // index of packet in stream
   unsigned int checksum;
   StreamMessage *next; // for linked list of early packets
   char data[1];
